Disentangle the ATs from the inner wiring of the Case API
